refactor: 🚧 works on rewrite backend

+++ src/utilities/db/Birthday.ts
@@ -5,48 +5,44 @@ import type { User } from 'discord.js';
 
 export class Birthday extends Utility {
 	public get = {
-		BirthdaysByDate: (date: Dayjs) =>
-			this.prisma.birthday.findMany({ where: { birthday: { contains: date.format('-MM-DD') } } }),
+		BirthdaysByDate: (date: Dayjs) => this.prisma.birthday.findMany({ where: { birthday: { contains: date.format('-MM-DD') } } }),
 		BirthdayByDateAndTimezone: (date: Dayjs, timezone: number) =>
 			this.prisma.birthday.findMany({
-				where: { birthday: { contains: date.format('-MM-DD') }, guild: { timezone } },
+				where: { birthday: { contains: date.format('-MM-DD') }, guild: { timezone } }
 			}),
 		BirthdayByDateTimezoneAndGuild: (date: Dayjs, timezone: number, guildId: string) => {
 			return this.prisma.birthday.findMany({
-				where: { birthday: { contains: date.format('-MM-DD') }, guild: { timezone, id: guildId } },
+				where: { birthday: { contains: date.format('-MM-DD') }, guild: { timezone, id: guildId } }
 			});
 		},
 		BirthdaysByGuildId: (guildId: string) => this.prisma.birthday.findMany({ where: { guildId } }),
 		BirthdayByUserAndGuild: (guildId: string, userId: string) =>
 			this.prisma.birthday.findUnique({
-				where: { userId_guildId: { guildId, userId } },
+				where: { userId_guildId: { guildId, userId } }
 			}),
-		BirthdaysNotDisabled: (guildId: string) =>
-			this.prisma.birthday.findMany({ where: { guildId, inDeleteQueue: false } }),
-		BirthdayCountByGuildId: (guildId: string) =>
-			this.prisma.birthday.count({ where: { guildId, inDeleteQueue: false } }),
+		BirthdaysNotDisabled: (guildId: string) => this.prisma.birthday.findMany({ where: { guildId, inDeleteQueue: false } }),
+		BirthdayCountByGuildId: (guildId: string) => this.prisma.birthday.count({ where: { guildId, inDeleteQueue: false } }),
 		BirthdayAvailableCount: () => this.prisma.birthday.count({ where: { inDeleteQueue: false } }),
-		BirthdayNotAvailableCount: () => this.prisma.birthday.count(),
+		BirthdayNotAvailableCount: () => this.prisma.birthday.count()
 	};
 
 	public update = {
 		BirthdayDisabled: (guildId: string, userId: string, inDeleteQueue: boolean) =>
 			this.prisma.birthday.update({
 				where: { userId_guildId: { guildId, userId } },
-				data: { inDeleteQueue },
+				data: { inDeleteQueue }
 			}),
 		BirthdayByUserAndGuild: (guildId: string, userId: string, birthday: string) =>
 			this.prisma.birthday.update({
 				where: { userId_guildId: { guildId, userId } },
-				data: { birthday },
-			}),
+				data: { birthday }
+			})
 	};
 
 	public delete = {
 		GuildById: (id: string) => this.prisma.guild.delete({ where: { id } }),
 		ByDisabledGuilds: () => this.prisma.guild.deleteMany({ where: { inDeleteQueue: true } }),
-		ByGuildAndUser: (guildId: string, userId: string) =>
-			this.prisma.birthday.delete({ where: { userId_guildId: { guildId, userId } } }),
+		ByGuildAndUser: (guildId: string, userId: string) => this.prisma.birthday.delete({ where: { userId_guildId: { guildId, userId } } })
 	};
 
 	private prisma = container.prisma;
@@ -54,7 +50,7 @@ export class Birthday extends Utility {
 	public constructor(context: Utility.Context, options: Utility.Options) {
 		super(context, {
 			...options,
-			name: 'birthday',
+			name: 'birthday'
 		});
 	}
 
@@ -65,21 +61,21 @@ export class Birthday extends Utility {
 				guild: {
 					connectOrCreate: {
 						create: { id },
-						where: { id },
-					},
+						where: { id }
+					}
 				},
 				user: {
 					connectOrCreate: {
 						create: {
 							id: user.id,
 							discriminator: user.discriminator,
-							username: user.username,
+							username: user.username
 						},
 						where: {
-							id: user.id,
-						},
-					},
-				},
-			},
+							id: user.id
+						}
+					}
+				}
+			}
 		});
 }
